backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-PyMongo==2.3.0
pymongo==4.6.0
python-dotenv==1.0.0
PyJWT==2.8.0
bcrypt==4.1.2
Werkzeug==3.0.1
gunicorn==21.2.0

# Analysis.py dependencies
python-docx==1.1.0
mediapipe==0.10.9
opencv-python==4.9.0.80
transformers==4.36.2
moviepy==1.0.3
librosa==0.10.1
matplotlib==3.8.2
numpy==1.24.3
scipy==1.11.4

# Interview.py dependencies
google-generativeai==0.3.2
pdfplumber==0.10.3
gTTS==2.5.0
pygame==2.5.2
SpeechRecognition==3.10.1
Pillow==10.2.0

# Resume.py dependencies
pandas==2.1.4
openpyxl==3.1.2

# Additional dependencies
python-multipart==0.0.6

flask
flask-cors
pymongo
python-dotenv
bcrypt
moviepy
opencv-python
mediapipe
numpy
librosa
transformers
python-docx
pdfplumber
google-generativeai
pandas
matplotlib
docx2pdf
speechrecognition
gTTS
pygame
pillow
PyJWT
werkzeug
